Add files via upload
Mesoscopic physics is one of those branches of modern condensed matter
physics that are developing rapidly. The progress achieved in the last few
decades in the manufacture of micron and sub-micron-sized samples has made
it possible to discover a whole range of new physical effects that are absent
in macroscopic samples and has made it possible to create solid-state devices
whose operating principle is based on quantum laws.
These lecture notes contain the basics of the theory of mesoscopic systems.
The presented theory allows us to describe the physics of the main phenom-
ena and the conditions for their observation in such systems. A characteristic
feature of a mesoscopic system is that its properties are determined by the
behaviour of a single quantum particle. Therefore, preserving the phase coher-
ence, spectrum quantization, and charge quantization are the components that
determine the occurrence of mesoscopic effects.
The lecture notes are intended for students and postgraduates who are
specializing in condensed matter physics, microelectronics, and nano physics.
The knowledge of the basics of quantum mechanics and statistical physics is
essential.